requirements.txt
asgiref==3.3.4
Django==3.2
django-dotenv==1.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
future==0.18.2
generics==3.3.1
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
serializers==0.2.4
sqlparse==0.4.1
typing-extensions==3.7.4.3
